move the files to their directories
