+++ translation/src/components/MainRow.css
