+++ translation/src/components/leftMenu/ProgressBar.jsx
