+++ translation/src/components/leftMenu/progressBar.css
