+++ translation/src/components/mainMenu/searchBar/Filter.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import './MainRow.css'
+import '../../../components/MainRow.css'
 import FilterAltIcon from '@mui/icons-material/FilterAlt';
 
 const Filter = ({data,setData,changeDataByContainerName}) => {
