+++ translation/src/components/mainMenu/searchBar/SearchBar.jsx
@@ -1,5 +1,6 @@
 import React, { useState, useEffect } from "react";
-import "./SearchBar.css";
+import '../../../components/MainRow.css'
+import '../searchBar/style/SearchBar.css'
 import SearchIcon from '@mui/icons-material/Search';
 import { Button } from "@mui/material";
