+++ translation/src/components/mainMenu/searchBar/SortBy.jsx
@@ -1,5 +1,5 @@
 import React, {useState} from "react";
-import './MainRow.css'
+import '../../../../src/'
 import InputLabel from '@mui/material/InputLabel';
 import MenuItem from '@mui/material/MenuItem';
 import FormControl from '@mui/material/FormControl';
