+++ translation/src/components/mainMenu/searchBar/style/SearchBar.css
